Feature/increase bullet velocity for the br55 ammo (#562)
* Ready to test

* Update changelog.md

* Update changelog.md


* Update changelog.md


---------
