+++ addons/overrides/data/weapons/config_weapons.hpp
@@ -1,4 +1,3 @@
-#include "br55\config_weapons.hpp"
 #include "bulldog\config_weapons.hpp"
 #include "grl45\config_weapons.hpp"
 #include "m7\config_weapons.hpp"
