+++ changelog.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Updated
+- Velocity for BR ammunition increased to improve effectiveness over distance.
 ### Fixed
 - Scorpion hidden texture filepath updated to correct path.
